[202406.xx] address error handling in PKCS#11 for multiple TLS connection use case
Due to possibility of user opening multiple TLS connections, the previous fix for memory leak is reverted (commented out) as it would cause all current TLS sessions to become invalid if any single connection errors out to TLS_Cleanup.

Instead, if an error occurs, only close the session so that new connections may create new sessions

The code is left as reference should de-initialization of the library be required.