[202406.xx] fix sockets not being freed properly when an unplanned disconnection happens
When an unplanned disconnection happens, the socket will first be shutdown, then closed.
However, in the lwip stack a disconnected socket will have its lwip_sock *sock->conn field as NULL, this will cause lwip_shutdown to return internal error and skip SOCKETS_Close
As a result, the remainder socket structures are also not freed and sockets_allocated is never released

This commit marks the unplanned disconnection case as valid, allowing SOCKETS_Close to be run and release the socket object completely.
In the case of unplanned disconnection, the lwip stack has already taken care of the lwip-related structs with done_socket(), so there should not be any memory leaks